+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 
 ### Fixed
 - Median statistics
+- Complex power of negative real numbers
 
 ### Security